refactor(getPageContentText): Relocate getPageContentText to a dedicated file and eliminate the redundant isIterable function.
