+++ lib/plugins/algolia.js
@@ -1,6 +1,6 @@
 import BLOG from '@/blog.config'
-import { getPageContentText } from '@/pages/search/[keyword]'
 import algoliasearch from 'algoliasearch'
+import { getPageContentText } from '@/lib/notion/getPageContentText'
 
 /**
  * 生成全文索引
