+++ lib/utils/post.js
@@ -6,11 +6,11 @@ import { getPostBlocks } from '@/lib/db/getSiteData'
 import { getPageTableOfContents } from '@/lib/notion/getPageTableOfContents'
 import { siteConfig } from '@/lib/config'
 import { getDataFromCache, setDataToCache } from '@/lib/cache/cache_manager'
-import { getPageContentText } from '@/pages/search/[keyword]'
 import { getAiSummary } from '@/lib/plugins/aiSummary'
 import BLOG from '@/blog.config'
 import { uploadDataToAlgolia } from '@/lib/plugins/algolia'
 import { countWords } from '@/lib/plugins/wordCount'
+import { getPageContentText } from '@/lib/notion/getPageContentText'
 
 /**
  * 获取文章的关联推荐文章列表，目前根据标签关联性筛选
